requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
python-dotenv==1.0.1
email-validator==2.2.0
PyJWT==2.10.1
Werkzeug==3.0.1
cloudinary==1.39.1
psycopg2-binary==2.9.9
requests==2.31.0